fix: trajectoryId and trajectoryName
