+++ src/gitInfo.ts
@@ -1,5 +1,5 @@
 export const GIT_INFO = {
   branch: "fix/ANT-3656_warning_modify_how_warning_are_retrieved",
-  commit: "39d1af3",
-  commitTime: "2025-07-25T14:39:41.270Z"
+  commit: "2773679",
+  commitTime: "2025-07-28T07:35:36.509Z"
 };
