+++ src/mocks/data/tests/warning.mock.ts
@@ -11,7 +11,7 @@ export const mockSingleWarningMessages: WarningMessage = {
   generatedBy: 'unknown_user',
   generatedAt: '2025-04-01T18:31:53.623683' as unknown as Date,
   trajectoryId: 105,
-  trajectory: 'areas_BP23_A_ref',
+  trajectoryName: 'areas_BP23_A_ref',
   secondTrajectory: 'links_BP23_A_ref',
   isAck: false,
 };
@@ -26,7 +26,7 @@ export const mockWarningMessagesWithTwo: WarningMessage[] = [
     generatedBy: 'unknown_user',
     generatedAt: '2025-04-01T18:31:53.623683' as unknown as Date,
     trajectoryId: 104,
-    trajectory: 'areas_BP23_A_ref',
+    trajectoryName: 'areas_BP23_A_ref',
     secondTrajectory: 'links_BP23_A_ref',
     isAck: false,
   },
@@ -38,30 +38,9 @@ export const mockWarningMessagesWithTwo: WarningMessage[] = [
     generatedBy: 'unknown_user',
     generatedAt: '2025-04-01T18:31:53.623683' as unknown as Date,
     trajectoryId: 104,
-    trajectory: 'areas_BP23_A_ref',
+    trajectoryName: 'areas_BP23_A_ref',
     secondTrajectory: 'links_BP23_A_ref',
     isAck: false,
-  },
-];
-
-export const mockTrajectoryTwo = [
-  {
-    id: 1,
-    trajectoryName: 'area_PB_2024',
-    type: TRAJECTORY_TYPE.AREA,
-    version: 3,
-    userName: 'mouad',
-    creationDate: '2024-07-22 15:13:56.860045' as unknown as Date,
-    loadArea: 'AT',
-  },
-  {
-    id: 2,
-    trajectoryName: 'area_PB_2026',
-    type: TRAJECTORY_TYPE.AREA,
-    version: 3,
-    userName: 'mouad',
-    creationDate: '2026-08-22 15:13:56.860045' as unknown as Date,
-    loadArea: 'BE',
   },
 ];
 
@@ -75,7 +54,7 @@ export const mockWarningMessages: WarningMessage[] = [
     generatedBy: 'unknown_user',
     generatedAt: '2025-04-01T18:31:53.623683' as unknown as Date,
     trajectoryId: 105,
-    trajectory: 'areas_BP23_A_ref',
+    trajectoryName: 'areas_BP23_A_ref',
     secondTrajectory: 'links_BP23_A_ref',
     isAck: false,
   },
@@ -87,7 +66,7 @@ export const mockWarningMessages: WarningMessage[] = [
     generatedBy: 'unknown_user',
     generatedAt: '2025-04-01T18:31:53.623683' as unknown as Date,
     trajectoryId: 105,
-    trajectory: 'areas_BP23_A_ref',
+    trajectoryName: 'areas_BP23_A_ref',
     secondTrajectory: 'links_BP23_A_ref',
     isAck: false,
   },
@@ -99,7 +78,7 @@ export const mockWarningMessages: WarningMessage[] = [
     generatedBy: 'unknown_user',
     generatedAt: '2025-04-01T18:31:53.623683' as unknown as Date,
     trajectoryId: 105,
-    trajectory: 'areas_BP23_A_ref',
+    trajectoryName: 'areas_BP23_A_ref',
     secondTrajectory: 'links_BP23_A_ref',
     isAck: false,
   },
@@ -111,7 +90,7 @@ export const mockWarningMessages: WarningMessage[] = [
     generatedBy: 'unknown_user',
     generatedAt: '2025-04-01T18:31:53.623683' as unknown as Date,
     trajectoryId: 105,
-    trajectory: 'areas_BP23_A_ref',
+    trajectoryName: 'areas_BP23_A_ref',
     secondTrajectory: 'links_BP23_A_ref',
     isAck: false,
   },
@@ -124,7 +103,7 @@ export const mockWarningMessages: WarningMessage[] = [
     generatedBy: 'unknown_user',
     generatedAt: '2025-04-01T18:31:53.623683' as unknown as Date,
     trajectoryId: 105,
-    trajectory: 'areas_BP23_A_ref',
+    trajectoryName: 'areas_BP23_A_ref',
     secondTrajectory: 'links_BP23_A_ref',
     isAck: false,
   },
@@ -136,7 +115,7 @@ export const mockWarningMessages: WarningMessage[] = [
     generatedBy: 'unknown_user',
     generatedAt: '2025-04-01T18:31:53.623683' as unknown as Date,
     trajectoryId: 105,
-    trajectory: 'areas_BP23_A_ref',
+    trajectoryName: 'areas_BP23_A_ref',
     secondTrajectory: 'links_BP23_A_ref',
     isAck: false,
   },
@@ -148,7 +127,7 @@ export const mockWarningMessages: WarningMessage[] = [
     generatedBy: 'unknown_user',
     generatedAt: '2025-04-01T18:31:53.623683' as unknown as Date,
     trajectoryId: 105,
-    trajectory: 'areas_BP23_A_ref',
+    trajectoryName: 'areas_BP23_A_ref',
     secondTrajectory: 'links_BP23_A_ref',
     isAck: false,
   },
@@ -160,7 +139,7 @@ export const mockWarningMessages: WarningMessage[] = [
     generatedBy: 'unknown_user',
     generatedAt: '2025-04-01T18:31:53.623683' as unknown as Date,
     trajectoryId: 105,
-    trajectory: 'areas_BP23_A_ref',
+    trajectoryName: 'areas_BP23_A_ref',
     secondTrajectory: 'links_BP23_A_ref',
     isAck: false,
   },
@@ -168,7 +147,6 @@ export const mockWarningMessages: WarningMessage[] = [
 
 export const mockDataMessage: DataWarningMessage = {
   ...mockSingleWarningMessages,
-  trajectory: 'load_BP_23_REF',
   trajectoryType: TRAJECTORY_TYPE.LOAD,
   onClickItem: null,
 };
