+++ src/shared/types/Warning.type.ts
@@ -12,8 +12,8 @@ export interface WarningMessage {
   code: string;
   generatedBy: string;
   generatedAt: Date;
-  trajectoryId: number | null;
-  trajectory: string;
+  trajectoryId: number;
+  trajectoryName: string;
   secondTrajectory: string;
   isAck: boolean;
 }
@@ -32,7 +32,7 @@ export interface DataWarningMessage extends WarningMessage {
 
 export interface CardDataType
   extends WithNullableFields<
-    Omit<WarningMessage, 'level' | 'trajectory' | 'secondTrajectory'>,
+    Omit<WarningMessage, 'level' | 'trajectoryId' | 'trajectoryName' | 'secondTrajectory'>,
     'id' | 'content' | 'generatedBy' | 'generatedAt'
   > {
   trajectoryId: number | null;
