+++ src/shared/utils/test/warningUtils.test.ts
@@ -16,7 +16,7 @@ describe('sortByLevel', () => {
         generatedBy: 'unknown_user',
         generatedAt: '2025-04-01T18:31:53.623683' as unknown as Date,
         trajectoryId: 105,
-        trajectory: 'areas_BP23_A_ref',
+        trajectoryName: 'areas_BP23_A_ref',
         secondTrajectory: 'links_BP23_A_ref',
         isAck: false,
       },
@@ -28,7 +28,7 @@ describe('sortByLevel', () => {
         generatedBy: 'unknown_user',
         generatedAt: '2025-04-01T18:31:53.623683' as unknown as Date,
         trajectoryId: 105,
-        trajectory: 'areas_BP23_A_ref',
+        trajectoryName: 'areas_BP23_A_ref',
         secondTrajectory: 'links_BP23_A_ref',
         isAck: false,
       },
@@ -41,7 +41,7 @@ describe('sortByLevel', () => {
         generatedBy: 'unknown_user',
         generatedAt: '2025-04-01T18:31:53.623683' as unknown as Date,
         trajectoryId: 105,
-        trajectory: 'areas_BP23_A_ref',
+        trajectoryName: 'areas_BP23_A_ref',
         secondTrajectory: 'links_BP23_A_ref',
         isAck: false,
       },
@@ -53,7 +53,7 @@ describe('sortByLevel', () => {
         generatedBy: 'unknown_user',
         generatedAt: '2025-04-01T18:31:53.623683' as unknown as Date,
         trajectoryId: 105,
-        trajectory: 'areas_BP23_A_ref',
+        trajectoryName: 'areas_BP23_A_ref',
         secondTrajectory: 'links_BP23_A_ref',
         isAck: false,
       },
@@ -65,7 +65,7 @@ describe('sortByLevel', () => {
         generatedBy: 'unknown_user',
         generatedAt: '2025-04-01T18:31:53.623683' as unknown as Date,
         trajectoryId: 105,
-        trajectory: 'areas_BP23_A_ref',
+        trajectoryName: 'areas_BP23_A_ref',
         secondTrajectory: 'links_BP23_A_ref',
         isAck: false,
       },
@@ -78,7 +78,7 @@ describe('sortByLevel', () => {
         generatedBy: 'unknown_user',
         generatedAt: '2025-04-01T18:31:53.623683' as unknown as Date,
         trajectoryId: 105,
-        trajectory: 'areas_BP23_A_ref',
+        trajectoryName: 'areas_BP23_A_ref',
         secondTrajectory: 'links_BP23_A_ref',
         isAck: false,
       },
@@ -90,7 +90,7 @@ describe('sortByLevel', () => {
         generatedBy: 'unknown_user',
         generatedAt: '2025-04-01T18:31:53.623683' as unknown as Date,
         trajectoryId: 105,
-        trajectory: 'areas_BP23_A_ref',
+        trajectoryName: 'areas_BP23_A_ref',
         secondTrajectory: 'links_BP23_A_ref',
         isAck: false,
       },
@@ -102,7 +102,7 @@ describe('sortByLevel', () => {
         generatedBy: 'unknown_user',
         generatedAt: '2025-04-01T18:31:53.623683' as unknown as Date,
         trajectoryId: 105,
-        trajectory: 'areas_BP23_A_ref',
+        trajectoryName: 'areas_BP23_A_ref',
         secondTrajectory: 'links_BP23_A_ref',
         isAck: false,
       },
